# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Community Quiz API",
        "description": "Community-Based Quiz Management System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/quizzes/{id}/share": {
            "get": {
                "tags": [
                    "Sharing"
                ],
                "summary": "Get a public shareable link for a quiz",
                "operationId": "shareQuizLink",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Share URL generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "quiz_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "share_url": {
                                            "type": "string",
                                            "example": "http://localhost:3000/quiz/1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quiz not found"
                    }
                }
            }
        },
        "/api/submissions/{id}/share": {
            "get": {
                "tags": [
                    "Sharing"
                ],
                "summary": "Get a shareable link for a quiz result",
                "operationId": "shareResultLink",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Submission ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Share URL generated successfully"
                    },
                    "404": {
                        "description": "Submission not found"
                    }
                }
            }
        },
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "List all user favorites",
                "operationId": "favoriteIndex",
                "responses": {
                    "200": {
                        "description": "Favorites retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Favorite"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Add a quiz or category to favorites",
                "operationId": "favoriteStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "target_type",
                                    "target_id"
                                ],
                                "properties": {
                                    "target_type": {
                                        "type": "string",
                                        "enum": [
                                            "quiz",
                                            "category"
                                        ]
                                    },
                                    "target_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Favorite saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Favorite"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/favorites/{favorite}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Remove an item from favorites",
                "operationId": "favoriteDestroy",
                "parameters": [
                    {
                        "name": "favorite",
                        "in": "path",
                        "description": "Favorite ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite removed successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Favorite not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Favorite": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "target_type": {
                        "type": "string",
                        "enum": [
                            "quiz",
                            "category"
                        ]
                    },
                    "target_id": {
                        "type": "integer"
                    },
                    "details": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}